services/ai-ml/requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-dotenv>=1.0.0
scikit-learn>=1.3.2
opencv-python>=4.9.0.80
numpy>=1.26.2
pandas>=2.1.4
pydantic>=2.5.3
httpx>=0.26.0
pillow>=10.0.0
